[indvars] Missing variables at Og:
https://bugs.llvm.org/show_bug.cgi?id=51735
https://github.com/llvm/llvm-project/issues/51077

Address @nikic comments about layering violation.
- Move code to 'LoopInfo.h'.
